resultHandler to hazardExport refactor
