+++ src/org/opensha2/DeaggCalc.java
@@ -8,7 +8,7 @@ import org.opensha2.calc.CalcConfig;
 import org.opensha2.calc.HazardCalcs;
 import org.opensha2.calc.Deaggregation;
 import org.opensha2.calc.Hazard;
-import org.opensha2.calc.ResultHandler;
+import org.opensha2.calc.HazardExport;
 import org.opensha2.calc.Site;
 import org.opensha2.calc.Sites;
 import org.opensha2.calc.ThreadCount;
@@ -149,7 +149,7 @@ public class DeaggCalc {
 
     log.info(PROGRAM + ": calculating ...");
 
-    ResultHandler handler = ResultHandler.create(config, sites, log);
+    HazardExport handler = HazardExport.create(config, sites, log);
 
     for (Site site : sites) {
       Hazard hazard = HazardCalc.calc(model, config, site, executor);
